added funtionality of score in views

+++ App_mobile/app/src/main/java/com/example/libreria_in_47_app/interfaces/IBook.java
@@ -12,6 +12,6 @@ public interface IBook {
     int getFormatId();
     int getEditorialId();
     int getCategoryId();
-    double getScore();
+    float getScore();
     int getNumberScores();
 }
